% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi-sqr-plot.R
\name{chi_pic}
\alias{chi_pic}
\title{Caculating and visualizing the correlation between two variables}
\usage{
chi_pic(dataset)
}
\arguments{
\item{dataset}{targeted data}
}
\value{
interactive correlation grah
}
\description{
Caculating and visualizing the correlation between two variables
}
\examples{
chi_pic(dataset)
}
\author{
Lin Quan
}
