% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{dataclean}
\alias{dataclean}
\title{read csv file from USDA site and clean up.}
\usage{
dataclean(url)
}
\arguments{
\item{url}{the csv file path.}
}
\value{
cleaned data.
}
\description{
read csv file from USDA site and clean up.
}
\examples{
file <-dataclean("https://www.ers.usda.gov/webdocs/DataFiles/48747/Unemployment.csv")
str(file)
}
\author{
Zirou Zhou
}
