% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{localunemployrate}
\alias{localunemployrate}
\title{read csv file from USDA site and clean up.}
\usage{
localunemployrate(file, local.name)
}
\arguments{
\item{file}{the file path.}

\item{local.name}{the chosen state.}
}
\value{
unemployment rate by year
}
\description{
read csv file from USDA site and clean up.
}
\examples{
localunemployrate(file, "Texas")
}
\author{
Zirou Zhou
}
