% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_hosehold_income.R
\name{plot_medianhouseholdincome}
\alias{plot_medianhouseholdincome}
\title{Plot the median house hold income in county level for a specific state in 2019}
\usage{
plot_medianhouseholdincome(file, yr = 2019, State.name)
}
\arguments{
\item{file}{the file path.}

\item{yr}{2019.}

\item{State.name}{the chosen state.}
}
\value{
a ggplot figure
}
\description{
Plot the median house hold income in county level for a specific state in 2019
}
\examples{
file=dataclean("https://www.ers.usda.gov/webdocs/DataFiles/48747/Unemployment.csv")

plotunemployed(file, 2019, "NJ")
}
\author{
Lin Quan
}
