% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianhhincome.R
\name{plotmedianhouseholdincome}
\alias{plotmedianhouseholdincome}
\title{Plot the 2019 median household income in county level for a specific state.}
\usage{
plotmedianhouseholdincome(file, State.name)
}
\arguments{
\item{file}{the file path.}

\item{State.name}{the selected state.}
}
\value{
a mapping figure
}
\description{
Plot the 2019 median household income in county level for a specific state.
}
\examples{
plotmedianhouseholdincome(file, "NJ")
}
\author{
Lin Quan
}
