% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment_rate.R
\name{plotunemployed}
\alias{plotunemployed}
\title{Plot the unemployment rate in county level for a specific state and a year}
\usage{
plotunemployed(file, yr, State.name)
}
\arguments{
\item{file}{the file path.}

\item{yr}{the select year.}

\item{State.name}{the chosen state.}
}
\value{
a ggplot figure
}
\description{
Plot the unemployment rate in county level for a specific state and a year
}
\examples{
# file=dataclean("https://www.ers.usda.gov/webdocs/DataFiles/48747/Unemployment.csv")

# plotunemployed(file, 2018, "NJ")
}
\author{
Xiaolan Wan
}
