% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment_rate_animation.R
\name{plotunemployed_animation}
\alias{plotunemployed_animation}
\title{Provide animation plot of the unemployment rate in county level for a specific state}
\usage{
plotunemployed_animation(file, State.name)
}
\arguments{
\item{file}{the file path.}

\item{State.name}{the chosen state.}
}
\value{
a gif figure
}
\description{
Provide animation plot of the unemployment rate in county level for a specific state
}
\examples{
file=dataclean("https://www.ers.usda.gov/webdocs/DataFiles/48747/Unemployment.csv")

plotunemployed_animation(file, "NJ")
}
\author{
Xiaolan Wan
}
