% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment_rate_time.R
\name{plotunemployed_time}
\alias{plotunemployed_time}
\title{Plot the unemployment rate along with years}
\usage{
plotunemployed_time(file, local.name)
}
\arguments{
\item{file}{the file path.}

\item{local.name}{the chosen local.}
}
\value{
a ggplot figure
}
\description{
Plot the unemployment rate along with years
}
\examples{
file=dataclean("https://www.ers.usda.gov/webdocs/DataFiles/48747/Unemployment.csv")

plotunemployed_time(file, "Texas")
}
\author{
Lin Quan Zirou Zhou
}
