% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{stateunemployed}
\alias{stateunemployed}
\title{Plot top 10 unemployed county histogram in selected state and a year}
\usage{
stateunemployed(file, yr, State.name)
}
\arguments{
\item{file}{the file path.}

\item{yr}{the select year.}

\item{State.name}{the chosen state.}
}
\value{
top 10 unemployed county histogram
}
\description{
Plot top 10 unemployed county histogram in selected state and a year
}
\examples{
file=dataclean("https://www.ers.usda.gov/webdocs/DataFiles/48747/Unemployment.csv")

stateunemployed(file, 2011, "IA")
}
\author{
Zirou Zhou Lin Quan
}
